+++ Corona Widget/Widget/View/LargeWidgetBlock.swift
@@ -56,11 +56,3 @@ struct countStack : View {
         } //: HSTACK
     } //: BODY
 }
-
-extension Double {
-    func withCommas() -> String {
-        let numberFormatter = NumberFormatter()
-        numberFormatter.numberStyle = .decimal
-        return numberFormatter.string(from: NSNumber(value:self)) ?? "\(Int(self))"
-    }
-}
